ms_predictor/requirements.txt
paho-mqtt
json
scikit-learn
numpy